updated dependencies + removed quartz

+++ src/main/kotlin/com/kraktun/kbot/commands/callbacks/CallbackChat.kt
@@ -3,4 +3,4 @@ package com.kraktun.kbot.commands.callbacks
 /**
  * Same as a Triple, but with significant names
  */
-data class CallbackChat(val callback: CallbackHolder, val user: Int, val chatInstance: String)
+data class CallbackChat(val callback: CallbackHolder, val user: Long, val chatInstance: String)
