+++ src/main/kotlin/com/kraktun/kbot/commands/core/MultiCommandChat.kt
@@ -1,3 +1,3 @@
 package com.kraktun.kbot.commands.core
 
-data class MultiCommandChat(val bot: String, val user: Int, val chat: Long)
+data class MultiCommandChat(val bot: String, val user: Long, val chat: Long)
